# [SpecDec] Remove Batch Expansion (2/3)