--- a/vllm/attention/backends/utils.py
+++ b/vllm/attention/backends/utils.py
@@ -313,7 +313,7 @@ class CommonAttentionState(AttentionState):
             seq_lens=None,
             seq_lens_tensor=self._graph_seq_lens[:batch_size],
             max_query_len=1,
-            decode_query_len=1,
+            max_decode_query_len=1,
             max_prefill_seq_len=0,
             max_decode_seq_len=self.runner.max_seq_len_to_capture,
             query_start_loc=None,