Follow up of #8839.

We will use `flash_attn_varlen_func` for MQA scorer. Therefore, we can support different propose lengths for different requests in the batch, which is essential for [ngram](https://github.com/vllm-project/vllm/blob/main/vllm/spec_decode/ngram_worker.py) and [dynamic speculative decoding](https://github.com/vllm-project/vllm/issues/4565).

The following are some preliminary benchmark numbers with MQA scorer compared with batch expansion with/w/o cuda graph.

<img width="538" alt="Screenshot 2024-10-11 at 2 44 07 PM" src="https://github.com/user-attachments/assets/61f74966-708a-4f8b-81fd-c065aba3d6be">