@LiuXiaoxuanPKU 2024-10-12
> LGTM. What happen if we enable CUDA graph in this PR?

It will not work. During the graph capture time, we still go into the normal decoding path [here](https://github.com/vllm-project/vllm/blob/02a48498497b6e843a01efba332179dfd5510591/vllm/attention/backends/flash_attn.py#L738). However, during the spec decode, we go into the varlen kernel branch [here](https://github.com/vllm-project/vllm/blob/02a48498497b6e843a01efba332179dfd5510591/vllm/attention/backends/flash_attn.py#L722). Therefore, the graph captured cannot be used during the scoring phase.