@comaniac 2024-10-12
Yeah I know. I mean what users will see in this case and we should provide a proper error message.