@LiuXiaoxuanPKU 2024-10-12
> Yeah I know. I mean what users will see in this case and we should provide a proper error message.

It will fall back to batch expansion if cuda graph is enabled as shown [here](https://github.com/vllm-project/vllm/blob/3cc9d84d43171b8319838cf9a0ba8f1c6f8299d3/vllm/spec_decode/spec_decode_worker.py#L200).