@wooyeonlee0 2024-10-21
Great! Thanks for your work @LiuXiaoxuanPKU 
I've been really looking forward to this feature :)
After testing the MQA feature, I got unexpected results that were somewhat different from yours.
In my experiment setting, the MQA scorer does not perform faster than batch expansion without cuda-graph.

Below I've shared the details of my experiment, which is quite similar to yours.
I hope sharing this helps you to build a complete MQA scorer :)

=========
GPU: 1 H100 (tp=1)
Model: meta-llama/Llama-2-7b-chat-hf (draft: JackFram/llama-68m)
input_len: 550
output_len: 150
num_speculative_tokens: 3
temperature: 0
script: benchmark_latency.py
vllm version: v0.6.3

**scoring time**
|batch_size|batch expansion|batch expansion w/o cuda graph|mqa scorer w/o cuda graph|
|--|--|--|--|
|1|7.20 ms|9.88 ms|10.19 ms|

**Avg latency**
|batch_size|batch expansion|batch expansion w/o cuda graph|mqa scorer w/o cuda graph|
|--|--|--|--|
|1|0.91 ms|1.34 ms|1.35 ms|