@LiuXiaoxuanPKU 2024-10-21
@wooyeonlee0 
Thanks for the attention! yeah there are some errors in the previous benchmark results. I just updated the numbers.

For the last column MQA with  cuda graph, we implemented a quick version (without handling edge cases) to test the performance. As seen from the numbers, we are thinking maybe instead of adding cuda graph support for MQA, which might introduce memory overhead because of the change of bucketing strategy, it might be easier to switch between MQA scorer and batch expansion scroer. For example, when batch_size < 8, use batch expansion scorer with cuda graph, when batch_size > 8, use MQA scorer without cudagraph. The switch should be relatively easy to implement without introducing overhead.

Any thoughts/comments/discussion are appreciated!