@wooyeonlee0 2024-10-23
@LiuXiaoxuanPKU Thanks for the quick response! :)

> adding cuda graph support for MQA, which might introduce memory overhead because of the change of bucketing strategy

Does the memory overhead come from multiple cuda graphs for each different K value?
Then how about to implement the other version of MQA scorer that works with a fixed K value, as the current batch expansion also supports fixed K value? (I'm not sure it's possible.)
This fixed-K MQA scorer might support cuda graph without such overhead, so it can completely replace the batch expansion.

I think we can use the variable-K MQA scorer only when we want to use the dynamic speculative decoding feature.
What do you think?