@bettybaii 2024-10-25
Thanks for your work [@LiuXiaoxuanPKU](https://github.com/LiuXiaoxuanPKU)
When I disabled CUDA Graphs (by adding the --enforce-eager parameter), I observed a significant increase in the overhead of token proposals. Below, I have provided the details of my experiment：）

GPU: 1 A10 (tp=1)
Model: Meta-Llama-3-8B (draft: turboderp/Qwama-0.5B-Instruct)
input_len: 256
output_len: 10
batch_size: 16
num_speculative_tokens: 3
temperature: 0
script: benchmark_throughput.py
vllm version: v0.6.3.post2

When I did not add the --enforce-eager parameter (automatically using the batch expansion scorer):
> INFO 10-25 02:19:22 spec_decode_worker.py:976] SpecDecodeWorker stage times: average_time_per_proposal_tok_ms=5.56 scoring_time_ms=43.28 verification_time_ms=1.52

When I added the --enforce-eager parameter (automatically using the MQA scorer):
> INFO 10-25 02:26:19 spec_decode_worker.py:976] SpecDecodeWorker stage times: average_time_per_proposal_tok_ms=13.69 scoring_time_ms=41.28 verification_time_ms=1.49


It can be observed that with CUDA Graphs enabled, the average_time_per_proposal_tok_ms is only 5.56 ms, whereas, with CUDA Graphs disabled, it increases to 13.69 ms :(

In addition, I conducted separate tests on the inference performance of the draft model with and without CUDA Graphs enabled, as shown below:
with CUDA Graphs enabled: TPOP = 5.307 ms
without CUDA Graphs enabled: TPOP = 14.780 ms

Why does enabling or disabling CUDA Graphs have such a significant impact on the draft model in my environment? Are there any solutions to address this issue?