@LiuXiaoxuanPKU 2024-11-06
> @LiuXiaoxuanPKU Thanks for the quick response! :)
> 
> > adding cuda graph support for MQA, which might introduce memory overhead because of the change of bucketing strategy
> 
> Does the memory overhead come from multiple cuda graphs for each different K value? Then how about to implement the other version of MQA scorer that works with a fixed K value, as the current batch expansion also supports fixed K value? (I'm not sure it's possible.) This fixed-K MQA scorer would support cuda graph without such memory overhead, so it can completely replace the batch expansion.
> 
> I think we can use the variable-K MQA scorer in this PR only when we want to use the dynamic speculative decoding feature. What do you think?

The tricky part is that in ngram, it very likely that requests within the same batch might have different propose lengths (0 or k). If there is a match for a given request, the propose length will be k, otherwise 0. In this case, we cannot assume the same k for all requests in the batch.